feat: move history to collapsible bottom bar on biomon
