+++ src/components/charsheet/ArmorView.tsx
@@ -1,11 +1,9 @@
-import { DamageHistoryPanel } from "../armor/DamageHistoryPanel";
 import { InventoryPanel } from "../armor/InventoryPanel";
 import { ShopPanel } from "../armor/ShopPanel";
 
 export const ArmorView = () => {
   return (
     <div class="container">
-      <DamageHistoryPanel />
       <InventoryPanel />
       <ShopPanel />
     </div>
